backend/services/decision_engine_fastapi/requirements.txt
fastapi
uvicorn
requests
pydantic
pydantic-settings
pyjwt
